fix: resolve warnings-as-errors gate breaks introduced by dispatch
T-S1-1's luminumbra_configure_first_party_target applied -Wall -Wextra
-Wpedantic -Werror -Wno-error=unused-parameter to first-party targets.
That gate caught real pre-existing bugs and several dispatch-introduced
issues. Net build state went from green to red; this commit restores
green by fixing each failure rather than relaxing the gate.

Fixes:
- WaterSystem.cpp:73 — entt view has no `.empty()` method on this
  template; replace with `begin() != end()` (introduced by
  T-DR-watersystem-camera-handle-fix).
- AudioPropagationSystem.h:11 — missing <condition_variable> and
  <queue> includes for the new propagation worker; pre-existing latent
  bug that surfaced because the file is now compiled (sources.cmake
  reconciled by T-S1-5).
- PhysicsSystem.cpp:454 — unused `to_listener` reflection vector; kept
  intent in a comment for the future angle-of-incidence weighting and
  suppressed the warning via (void)glm::normalize(...).
- SHIELD_WorldSystem.cpp:43 — initializer order mismatched declaration
  order (m_water_system declared after m_params but initialized first).
  Reorder the init list to match the header.
- Debug.h:17 — broken `#include "Log.h"` for client-side asserts;
  qualified to `luminumbra_common/core/Log.h` to match the actual
  Log.h location (pre-existing latent bug).
- main_client.cpp:212/215/233 — int/size_t comparison (`int` g_index vs
  `size_t` vector size); cast g_index to size_t at compare sites.

Also raises quality_gate.min_calibration_samples to 30 in
.forge/config.yaml. The default 10 was tuned for the EMA calibrator;
the current median calibrator needs more samples to settle on
heterogeneous task batches (the first dispatch learned a 10.5x
multiplier from 18 samples and rejected 16 of 16 subsequent tasks).

Test results after fixes: 18 of 23 pass, including all 3 SHIELD C0
fixes (SurfaceIsGeneratedAtCorrectHeight, NormalsPointUpwardsOnHori-
zontalSurface, CavesChangeGeneratedMesh), the new LOD-winding
regression, asset-processor round-trip, and meshing-skips-Loading
tests. SdfGpuCpuParityTest skips (no GL context in test binary).

The 5 failing tests are SEGFAULTs in the new
WorldStreamingStateStressTest and JobSystemStressTest written by
T-S2-jobsystem-stress-tests. They exercise the chunk-streaming and
JobSystem-ownership code introduced by T-S2-1 and T-S2-4a-d under
heavy concurrency and find real race conditions. These are real bugs
in the dispatch's new code; deferred to a follow-up.

+++ src/luminumbra_client/main_client.cpp
@@ -209,10 +209,10 @@ int main(int argc, char* argv[]) {
             case GameState::WORLD_LOADING:
             {
                 const int JOBS_PER_FRAME = 64; // Dispatch more jobs per frame for a faster visual fill
-                if (g_generation_dispatch_index < g_initial_chunks_to_load.size()) {
+                if (static_cast<size_t>(g_generation_dispatch_index) < g_initial_chunks_to_load.size()) {
                     std::vector<Luminumbra::IVec3> batch_to_generate;
                     int dispatched_this_frame = 0;
-                    while (g_generation_dispatch_index < g_initial_chunks_to_load.size() && dispatched_this_frame < JOBS_PER_FRAME) {
+                    while (static_cast<size_t>(g_generation_dispatch_index) < g_initial_chunks_to_load.size() && dispatched_this_frame < JOBS_PER_FRAME) {
                         const auto& coords = g_initial_chunks_to_load[g_generation_dispatch_index];
                         batch_to_generate.push_back(coords);
                         g_loading_visualizer->UpdateChunkState(coords, Luminumbra::Client::ChunkLoadVisualState::DISPATCHED);
@@ -230,7 +230,7 @@ int main(int argc, char* argv[]) {
                 glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
                 g_loading_visualizer->UpdateAndRender(deltaTime, "CONSTRUCTING WORLD GEOMETRY...", progress);
 
-                if (g_generation_dispatch_index >= g_initial_chunks_to_load.size()) {
+                if (static_cast<size_t>(g_generation_dispatch_index) >= g_initial_chunks_to_load.size()) {
                     LUMINUMBRA_CORE_INFO("World generation phase complete. Entering world.");
                     audioManager->StopMusic();
                     audioManager->PlayOneShot2D("ui_world_loaded"); // Play a sound on completion
